refactor: update module names and paths
