+++ nyx/moirai_ecs/component/nyx_component.py
